refactor: removes all todos and fixmes [skip ci]
