+++ src/components/molecules/Navbar.tsx
@@ -36,9 +36,6 @@ const Navbar: FC<NavbarProps> = () => {
             <NavLink text="Profile" />
           </LinkContainer>
         )}
-        {
-          //FIXME: adding a hidden navlink just to fill space does not feel like the best practice solution
-        }
         {!user && <NavLink text="Profile" style={{ opacity: 0 }} />}
       </Nav>
     </BSNavbar>
