+++ src/components/pages/LoginPage.tsx
@@ -1,14 +1,9 @@
 import LoginPageTemplate from 'components/templates/LoginPageTemplate';
-import { Contact, Message } from 'models';
-import { IContactParams } from 'models/Contact';
 import React, { FC, useContext, useEffect } from 'react';
 import { useHistory } from 'react-router';
-import { APP_ACTIONS } from 'store/App/appActions';
-import { USER_ACTIONS } from 'store/User/userActions';
-import { setupUser } from 'store/User/userController';
 import UserStore from 'store/User/UserStore';
+import { recoverUser } from 'store/User/userUtils';
 import LocalStorage from 'utils/LocalStorage';
-import { PeerId } from 'peer-id';
 
 // import Logger from 'utils/Logger';
 
@@ -42,37 +37,3 @@ const LoginPage: FC<LoginPageProps> = () => {
 };
 
 export default LoginPage;
-
-// TODO: not sure where to put this
-export interface IUserRecData {
-  keystore: PeerId;
-  contacts: Contact[];
-}
-export const recoverUser = async (userRecData: IUserRecData, dispatch: any) => {
-  const { keystore, contacts } = userRecData;
-  try {
-    const payload = await setupUser(
-      keystore,
-      (payload: { contact: Contact; message: Message }) => {
-        dispatch({
-          type: USER_ACTIONS.RECEIVE_MESSAGE,
-          payload,
-        });
-      },
-    );
-    dispatch({ type: USER_ACTIONS.SET_CLIENT, payload });
-    const newContacts = contacts
-      ? await Promise.all(
-          contacts.map((contact: IContactParams) => Contact.new(contact)),
-        )
-      : [];
-    dispatch({
-      type: USER_ACTIONS.SET_CONTACTS,
-      payload: {
-        contacts: newContacts,
-      },
-    });
-  } catch (err) {
-    return { type: APP_ACTIONS.SET_ERROR, payload: err };
-  }
-};
