+++ src/store/User/UserStore.tsx
@@ -7,7 +7,6 @@ import { UserAction } from './userActions';
 import Middleware from 'store/storeUtils/middleware';
 import { IState } from 'store/storeUtils/interfaces';
 
-//TODO: separate not-user-connection related stuff out
 export interface IUserState extends IState {
   user: User | undefined;
   clientNode?: libp2p;
